% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{Client}
\alias{Client}
\title{FHIR client connection}
\usage{
Client(
  app_id = NULL,
  app_secret = NULL,
  api_base,
  redirect_uri = NULL,
  patient_id = NULL,
  scope = list(),
  launch_token = NULL,
  state = NULL
)
}
\arguments{
\item{app_id}{The app ID.}

\item{app_secret}{The client secret.}

\item{api_base}{The FHIR service to connect.}

\item{redirect_uri}{The callback/redirect URL for your app.}

\item{patient_id}{The patient id against which to operate, if already known.}

\item{scope}{Space-separated list of scopes to request, if other than default.}

\item{launch_token}{The launch token.}

\item{state}{The client state.}
}
\description{
FHIR client connection
}
\examples{

settings = list('app_id'= 'my_web_app', 'api_base'= 'https://r4.smarthealthit.org')
smart <- Client("my_app", api_base = "https://r4.smarthealthit.org")

}
