% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{getModel}
\alias{getModel}
\title{getModel}
\usage{
getModel(model)
}
\arguments{
\item{model}{The resource model to retrieve}
}
\value{
A list of classes in the model
}
\description{
getModel
}
\examples{
getModel('patient')
}
