% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{getModelClass}
\alias{getModelClass}
\title{getModel classes}
\usage{
getModelClass(model)
}
\arguments{
\item{model}{The resource model to retrieve}
}
\value{
A list of classes in the model
}
\description{
getModel classes
}
\examples{
getModelClass('patient')
}
