% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{group_index}
\alias{group_index}
\title{Add group by}
\usage{
group_index(tibble, id, column = "name")
}
\arguments{
\item{tibble}{The \code{tibble} object to group}

\item{id}{The index id for each group}

\item{column}{The column to group.}
}
\description{
Add group by
}
