% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fhirclient.R
\name{install_fhirclient}
\alias{install_fhirclient}
\title{install fhirclient python package}
\usage{
install_fhirclient(
  envname = "fhirclient",
  method = "conda",
  version = "latest"
)
}
\arguments{
\item{envname}{Virtual environment to install the \code{fhirclient}
python module}

\item{method}{Use conda or virtualenv to install \code{fhirclient}.}

\item{version}{The version of \code{fhirclient}. By default, it will
install the latest version from the github master branch.}
}
\value{
install \code{fhirclient} python module
}
\description{
install fhirclient python package
}
