% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Model}
\alias{Model}
\title{Model}
\usage{
Model(model, class, jsonDict = list())
}
\arguments{
\item{model}{The resource model to build}

\item{class}{The class of the resource model}

\item{jsonDict}{The json list to build the data model object. If
NULL, then it returns the class.}
}
\value{
A \code{Resource} object
}
\description{
Model
}
\examples{
getModelClass('patient')
Model("patient", "Patient", list(id = "test"))
}
