% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.R
\name{Patient}
\alias{Patient}
\title{Patient class}
\usage{
Patient(jsonDict = list())
}
\arguments{
\item{jsonDict}{A list of items for the Patient resource.}
}
\value{
A object of Patient
}
\description{
Patient class
}
\examples{
Patient(list(id = "pt1"))
}
