% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Py-class}
\alias{Py-class}
\alias{$,Py-method}
\alias{$<-,Py-method}
\alias{names,Py-method}
\alias{py}
\title{Python object}
\usage{
\S4method{$}{Py}(x, name)

\S4method{$}{Py}(x, name) <- value

\S4method{names}{Py}(x)

py(x)
}
\arguments{
\item{x}{Resource object}

\item{name}{The class name to extract}

\item{value}{The value to assign}
}
\value{
python object
}
\description{
Python object

Extract input values by name

get py slot
}
