% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{Search}
\alias{Search}
\title{Search records}
\usage{
Search(
  ct,
  model,
  class,
  where = list(),
  include = NULL,
  page = c("first", "all"),
  tidy = TRUE
)
}
\arguments{
\item{ct}{The object from Client. The server to perform search.}

\item{model}{The resource model to build.}

\item{class}{The class of the resource model.}

\item{where}{The search parameters.}

\item{include}{The \code{include} parameter to include other resources.}
}
\description{
Search records
}
\examples{
\donotrun{
Search(ct, "observation", "Observation",
list(code = "http://loinc.org|2339-0"), include = "subject")
}
}
